requirements.txt
# LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-tavily>=0.2.0

# AI/ML
openai>=1.0.0
sentence-transformers>=2.0.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.4

# Hugging Face
langchain-huggingface>=0.1.0
transformers>=4.30.0

# PDF 처리
pypdf>=4.0.0
reportlab>=4.0.0

# 데이터 처리
pydantic>=2.0.0
python-dotenv>=1.0.0

# 유틸리티
tavily-python>=0.5.0

